Specification - Operator: Add helper functions `clauses/1` and `rule/1`
The functions return a result tuple containing the clauses or the rule
atom of the given operator.

In addition each function has a bang version which raises an
`ArgumentError` in case the given argument is not an operator.

Furthermore the `Operator` module now has a `operator?/1` function which
returns a boolean if the given value is an operator.