Specification: Mayor refactoring
This commit does a major refactoring of the core implementation of
Specification.

It mainly does two things, it ...

1. adds a `Result` struct
2. splits the evaluation into separate modules

`Result`
--------

The `Result` struct contains the following fields:

- `rule` The rule which was applied
- `result` The result produced by applying the value to the rule
- `value` The applied value

This struct can than be used to generate any kind of informative report.

Evaluation
----------

The evaluation has been split into modules based on the type of the
rule. Right now Specification supports this rule types:

- Function
- Module
- Operator
- Struct

For further details what that means take a look at the corresponding
@moduledocs.