Rule - Struct: Rewrite the struct rule and add a `NestedRule`
The `NestedRule` can, as the `Struct` rule can, be `use`d. It expects
two keys:

- `aggregator` the function to aggregate the evaluated nested rules
- `nested` the key to the `nested_rules`

Alternatively the using module itself can implement `aggregator/1`
and/or `nested_rules/1` in case it wants more control.

In turn an implementation for `Nestable` will be provided, as will an
`evaluate/2` function which takes all nested rules, aggregates them and
returns a result tuple based on the aggregation.