Rule - Nestable: Move everything related to Nestable to Operator
NestedStruct basically acted like a Operator and was also used to build
the built-in operators, as such it made sense to migrate all this logic
into `Spex.Operator` and as such make it straight forward to implement
custom operators.

Also rename it to `Aggregatable` and move the `NestedStruct` use logic
into a `Spex.Operator.Builder` module.

In the following step the `rules` references from NestedStruct will be
replaced with `clauses` to clean up old terminology.