Spec - Rule: Rewrite the ModuleSpec
It now checks that only modules are handled properly and atoms raise an
UndefinedFunctionError, as do modules without a `evaluate/1` function.

In addition the raise and throw tests have been rewritten to no longer
catch/rescue but instead simply crash.

The thought behind this, is that a module rule is a rather deliberately
written piece of code, compared to a function rule. If such a module
rule raises an error, or throws a term it is the responsibility of the
developer to handle this errors in his own module rule.

Function rules on the other hand might raise a `FunctionClauseError`
which does not necessarily mean negligence on the side of the developer
but simply is a limitation of the provided function at hand.

Spex will probably allow to configure the caught errors for function
rules or turn this feature of completely.