Result - Formatter: Fix code insertion in `before_compile`
To insert code the before compile hook as to be a macro - see `Module`
docs in the "Compile Callbacks" section.

Furthermore the `format/1` definition now takes a list and calls
`Enum.map` with `format/1` for each element.

Due to this being part of the before compile hook overridding this
defintion is trivial.